Align api/fixtures code platform wise
