+++ api/fixtures/example_none.go
@@ -1,0 +1,5 @@
+package fixtures
+
+type ExampleNoneOptions struct {
+	APIServerAddress string
+}
